gui-v2/backend/requirements.txt
# Modern FastAPI Backend Dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.35.0
python-socketio>=5.13.0
python-multipart>=0.0.20
websockets>=15.0.0

# Database and ORM
sqlalchemy>=2.0.25
alembic>=1.13.3
asyncpg>=0.29.0  # PostgreSQL async driver

# Authentication and Security
python-jose[cryptography]>=3.5.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.20

# Validation and Serialization
pydantic>=2.11.0
pydantic-settings>=2.6.1

# HTTP Client
httpx>=0.28.0

# Background Tasks
celery>=5.4.0
redis>=5.2.0

# Monitoring and Observability
structlog>=24.4.0
rich>=13.9.4
prometheus-client>=0.21.0

# Development
pytest>=8.3.4
pytest-asyncio>=0.24.0
black>=24.10.0
isort>=5.13.2
mypy>=1.13.0